# lib/api-zod/src/generated/types/sessionEnvelope.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YucaTanaTrades API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionState } from './sessionState';
import type { SessionUser } from './sessionUser';

export interface SessionEnvelope {
  state: SessionState;
  user: SessionUser | null;
  /** @nullable */
  expiresAt: Date | null;
  /**
     * Synchronizer token for unsafe requests in this session; never a session credential.
     * @minLength 32
     */
  csrfToken: string;
}