# lib/api-zod/src/generated/types/sessionUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YucaTanaTrades API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionUser {
  id: string;
  email: string;
  /** @nullable */
  displayName: string | null;
  emailVerified: boolean;
}