# lib/api-zod/src/generated/types/signInInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YucaTanaTrades API
 * OpenAPI spec version: 0.1.0
 */

export interface SignInInput {
  /** @maxLength 320 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 1024
     */
  password: string;
}